+++ src/v3/features/meeting_materials/public_talks/index.types.ts
@@ -1,0 +1,3 @@
+import { PublicTalksViewType } from '@definition/public_talks';
+
+export type PublicTalksType = { view: PublicTalksViewType };
