+++ src/v3/features/meeting_materials/public_talks/tableView/useTableView.tsx
@@ -1,0 +1,10 @@
+import { useRecoilValue } from 'recoil';
+import { publicTalksFilteredState } from '@states/publicTalks';
+
+const useTableView = () => {
+  const talksList = useRecoilValue(publicTalksFilteredState);
+
+  return { talksList };
+};
+
+export default useTableView;
