--- src/v3/features/public_talks/components/tableView/useTableView.tsx
@@ -1,10 +1,0 @@
-import { useRecoilValue } from 'recoil';
-import { publicTalksFilteredState } from '@states/publicTalks';
-
-const useTableView = () => {
-  const talksList = useRecoilValue(publicTalksFilteredState);
-
-  return { talksList };
-};
-
-export default useTableView;
