+++ src/v3/pages/meeting_materials/public_talks_list/usePublicTalksList.tsx
@@ -1,11 +1,9 @@
 import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
-import { congAccountConnectedState } from '@states/app';
 import { publicTalksState } from '@states/publicTalks';
 import { PublicTalksViewType } from '@definition/public_talks';
 
 const usePublicTalksList = () => {
-  const isConnected = useRecoilValue(congAccountConnectedState);
   const talksList = useRecoilValue(publicTalksState);
 
   const [currentView, setCurrentView] = useState<PublicTalksViewType>('list');
@@ -17,7 +15,7 @@ const usePublicTalksList = () => {
     });
   };
 
-  return { isConnected, talksList, handleToggleView, currentView };
+  return { talksList, handleToggleView, currentView };
 };
 
 export default usePublicTalksList;
