+++ src/v3/services/dexie/assignment.ts
@@ -1,5 +1,5 @@
 import { LANGUAGE_LIST } from '@constants/index';
-import { AssignmentCode } from '@definition/schedules';
+import { AssignmentCode } from '@definition/assignment';
 import { getTranslation } from '@services/i18n/translation';
 import appDb from '@shared/indexedDb/appDb';
