+++ src/v3/states/settings.ts
@@ -4,9 +4,9 @@ Individual property are evaluated using recoil selector
 */
 
 import { atom, selector } from 'recoil';
-import { SettingsType } from '@definition/app';
 import { AccountTypeState } from '@definition/api';
 import { TimeAwayType } from '@definition/person';
+import { SettingsType } from '@definition/settings';
 
 export const settingsState = atom({
   key: 'settings',
@@ -118,7 +118,7 @@ export const userAvatarState = selector({
   get: ({ get }) => {
     const settings = get(settingsState);
 
-    return settings.user_avatar || undefined;
+    return settings.user_avatar;
   },
 });
 
@@ -324,22 +324,6 @@ export const isMeetingEditorRoleState = selector({
   },
 });
 
-export const avatarUrlState = selector({
-  key: 'avatarUrl',
-  get: ({ get }) => {
-    const userAvatar = get(userAvatarState);
-
-    let src = '';
-
-    if (userAvatar && typeof userAvatar === 'object') {
-      const blob = new Blob([userAvatar]);
-      src = URL.createObjectURL(blob);
-    }
-
-    return src;
-  },
-});
-
 export const autoAssignMMOpeningPrayerState = selector({
   key: 'autoAssignMMOpeningPrayer',
   get: ({ get }) => {
@@ -400,14 +384,5 @@ export const userTimeAwayState = selector<TimeAwayType[]>({
     const settings = get(settingsState);
 
     return settings?.user_time_away || [];
-  },
-});
-
-export const publicTalkSyncState = selector({
-  key: 'publicTalkSync',
-  get: ({ get }) => {
-    const settings = get(settingsState);
-
-    return settings?.public_talk_sync || '';
   },
 });
